Added support for entity asset loading.

+++ solution/RenderEngine/include/AGraphicsWrapper.h
@@ -42,7 +42,7 @@ protected:
 
     virtual void setupBufferSubData(GUIMeshData* meshData) =0;
 
-    virtual void drawElement(uint32_t indicesId, uint32_t indicesSize) = 0;
+    virtual void drawElement(uint32_t indicesId, size_t indicesSize) = 0;
 
     virtual void disableVertexAttribute(int location) = 0;
     virtual void enableDepthTest() = 0;
