+++ solution/RenderEngine/include/OpenGLAPI.h
@@ -39,7 +39,7 @@ protected:
 
     void setupBufferSubData(GUIMeshData* meshData) override;
 
-    void drawElement(uint32_t indicesId, uint32_t indicesSize) override;
+    void drawElement(uint32_t indicesId, size_t indicesSize) override;
 
     void disableVertexAttribute(int location) override;
     void enableDepthTest() override;
